+++ test/learn_kit/naive_bayes/gaussian_test.exs
@@ -3,6 +3,10 @@ defmodule LearnKit.NaiveBayes.GaussianTest do
 
   alias LearnKit.NaiveBayes.Gaussian
 
+  setup_all do
+    {:ok, classificator: Gaussian.new([{:label1, [[-1, -1], [-2, -1], [-3, -2]]}, {:label2, [[1, 1], [2, 1], [3, 2], [-2, -2]]}])}
+  end
+
   test "create new knn classificator with empty data set" do
     assert %Gaussian{data_set: data_set} = Gaussian.new
 
@@ -18,9 +22,8 @@ defmodule LearnKit.NaiveBayes.GaussianTest do
     assert data_set == [b1: [[2, 3]], a1: [[1, 3], [1, 2]]]
   end
 
-  test "fit data set" do
-    classificator = LearnKit.NaiveBayes.Gaussian.new([{:label1, [[-1, -1], [-2, -1], [-3, -2]]}, {:label2, [[1, 1], [2, 1], [3, 2], [-2, -2]]}])
-    %LearnKit.NaiveBayes.Gaussian{fit_data: fit_data} = classificator |> LearnKit.NaiveBayes.Gaussian.fit
+  test "fit data set", state do
+    %Gaussian{fit_data: fit_data} = state[:classificator] |> Gaussian.fit
 
     assert fit_data == [
                         label1: [
@@ -34,27 +37,24 @@ defmodule LearnKit.NaiveBayes.GaussianTest do
                         ]
   end
 
-  test "return probability estimates for the feature" do
-    classificator = LearnKit.NaiveBayes.Gaussian.new([{:label1, [[-1, -1], [-2, -1], [-3, -2]]}, {:label2, [[1, 1], [2, 1], [3, 2], [-2, -2]]}])
-    classificator = classificator |> LearnKit.NaiveBayes.Gaussian.fit
+  test "return probability estimates for the feature", state do
+    classificator = state[:classificator] |> Gaussian.fit
 
-    assert {:ok, result} = classificator |> LearnKit.NaiveBayes.Gaussian.predict_proba([1, 2])
+    assert {:ok, result} = classificator |> Gaussian.predict_proba([1, 2])
     assert result == [label1: 0.0, label2: 0.017199571]
   end
 
-  test "return exact prediction for the feature" do
-    classificator = LearnKit.NaiveBayes.Gaussian.new([{:label1, [[-1, -1], [-2, -1], [-3, -2]]}, {:label2, [[1, 1], [2, 1], [3, 2], [-2, -2]]}])
-    classificator = classificator |> LearnKit.NaiveBayes.Gaussian.fit
+  test "return exact prediction for the feature", state do
+    classificator = state[:classificator] |> Gaussian.fit
 
-    assert {:ok, result} = classificator |> LearnKit.NaiveBayes.Gaussian.predict([1, 2])
+    assert {:ok, result} = classificator |> Gaussian.predict([1, 2])
     assert result == {:label2, 0.017199571}
   end
 
-  test "returns the mean accuracy on the given test data and labels" do
-    classificator = LearnKit.NaiveBayes.Gaussian.new([{:label1, [[-1, -1], [-2, -1], [-3, -2]]}, {:label2, [[1, 1], [2, 1], [3, 2], [-2, -2]]}])
-    classificator = classificator |> LearnKit.NaiveBayes.Gaussian.fit
+  test "returns the mean accuracy on the given test data and labels", state do
+    classificator = state[:classificator] |> Gaussian.fit
 
-    assert {:ok, result} = classificator |> LearnKit.NaiveBayes.Gaussian.score
+    assert {:ok, result} = classificator |> Gaussian.score
     assert result == 0.857143
   end
 end